Added more z80 instructions, and stack

+++ cpu/Z80Memory.cpp
@@ -2,25 +2,25 @@
 #include <fstream>
 #include <Z80Memory.h>
 using namespace std;
-uint8_t Z80Memory::fread8(int location)
+uint8_t Z80Memory::fread8(uint16_t location)
 {
 	return ram[location];
 } 
-uint8_t Z80Memory::fwrite8(int location,uint8_t data)
+uint8_t Z80Memory::fwrite8(uint16_t location,uint8_t data)
 {
 	ram[location]=data;
 	return 0;
 }
-uint16_t Z80Memory::fread16(int location)
+uint16_t Z80Memory::fread16(uint16_t location)
 {
 	return fread8(location) + (fread8(location + 1) << 8);
 }
-uint16_t Z80Memory::fwrite16(int location,uint16_t data)
+uint16_t Z80Memory::fwrite16(uint16_t location,uint16_t data)
 {
 	ram[location]=data;
 	return 0;
 }
-void Z80Memory::loadRom(string location, int start)
+void Z80Memory::loadRom(string location, uint16_t start)
 {
 	ifstream in(location.c_str());
 	unsigned char temp=0;
@@ -43,5 +43,4 @@ void Z80Memory::clear()
 		ram[i]=0;
 		i++;
 	}
-	//TODO: CLEAR MEMORY
 }
